VideoPress: Add raw caption on jetpack_videopress field (#26409)
* Set the video caption as the attachment excerpt

As per pe4Cmx-fL-p2#comment-153 comment, the caption for VideoPress attachments is stored on the excerpt field.

* Remove duplicated line

* Set the caption field on the returned VideoPress data for non-WPCOM sites

When we are not in WPCOM, the caption is fetched from video info returned by video_get_info_by_blogpostid.

* Set the caption field for WPCOM sites

The caption is stored on the excerpt field of the attachment post. On WPCOM, we need to fetch the post using the correct $blog_id.

This code is based on the `video_get_description` code, where the attachment post is fetched and then the content is read from it.

* Include changelog file

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3127877439