VideoPress: Expose playback token generation endpoint (#26974)
* Register route to generate VideoPress playback tokens

This route will be consumed mostly when the client needs to show a video or a thumbnail for a private video. In this case, if we do not provide a playback token, the request fails as a 403 response.

* Add changelog file

Co-authored-by: Damián Suárez <rdsuarez@gmail.com>

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/3297883836