VideoPress: Delay enabling "Done" button to mitigate failed video preview requests (#29493)
* Delay Done button activation to mitigate risk of race condition

The Done button appears after the upload of a new video is complete and the title + poster of the video are set.

By adding a delay after this process takes place, we give some time for the VideoPress backend to be ready to process a preview request from WordPress case the user saves the post right after clicking the Done button.

* Include debug messages to the button enabling delay

* Only set isProcessing to false if it's still true

Committed via a GitHub action: https://github.com/Automattic/jetpack/actions/runs/4440594485